Fix part_1

+++ part_1/right_triangle.rb
@@ -1,9 +1,9 @@
 puts "Введите сторону треугольника"
-x1 = gets.chomp.to_i
+x1 = gets.chomp.to_f
 puts "Введите сторону треугольника"
-x2 = gets.chomp.to_i
+x2 = gets.chomp.to_f
 puts "Введите сторону треугольника"
-x3 = gets.chomp.to_i
+x3 = gets.chomp.to_f
 a,b,c = [x1,x2,x3].sort
 if a == b && a == c && b == c
   puts "Треугольник равносторонний"
